Same for createQueryStr now in quellHelpers.test, both old test files removed
